requirements.txt
fastapi
uvicorn
python-multipart
pydantic
langchain
huggingface-hub
chromadb
PyMuPDF
python-dotenv
